Manually merged qasm_semantic.hpp.
